+++ authentication/tests.py
@@ -5,7 +5,7 @@ from django.core.urlresolvers import reverse
 from rest_framework import status
 from rest_framework.test import APITestCase
 
-from user_auth.models import ReporterProfile
+from authentication.models import ReporterProfile
 
 # Create your tests here.
